backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
faster-whisper>=1.0.3
groq>=0.11.0
python-dotenv>=1.0.0
setuptools<81

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Authentication
clerk-sdk-python>=0.1.0

# Email
sendgrid>=6.10.0

# Exports
reportlab>=4.0.0
python-docx>=1.1.0
jinja2>=3.1.0

# File Storage (AWS S3)
boto3>=1.34.0

# WebSockets for real-time
websockets>=12.0

# Additional utilities
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
